+++ testocache/cache.go
@@ -13,16 +13,18 @@
 package testocache
 
 import (
+	"bytes"
 	"cmp"
 	"errors"
 	"flag"
-	"io/fs"
+	"hash/fnv"
+	"io"
 	"os"
+	"path"
 	"path/filepath"
+	"slices"
 	"strconv"
-	"strings"
 	"sync"
-	"unicode"
 )
 
 var (
@@ -38,8 +40,17 @@ var (
 	)
 )
 
-// ErrDisabled indicates that caching is disabled.
-var ErrDisabled = errors.New("cache is disabled")
+var (
+	// ErrDisabled indicates that caching is disabled.
+	ErrDisabled = errors.New("testocache: cache is disabled")
+
+	// ErrInvalidKey indicates that passed key is invalid.
+	// Currently, key is invalid if it contains a NUL-byte.
+	ErrInvalidKey = errors.New("testocache: invalid key")
+
+	// ErrNotFound indicates that value was not found the passed key.
+	ErrNotFound = errors.New("testocache: not found")
+)
 
 const (
 	permFile os.FileMode = 0o600
@@ -56,10 +67,36 @@ func Disabled() bool {
 var kvMu sync.RWMutex
 
 // Keys returns all glob-matched keys by the given pattern.
-// E.g. "myplugin-prefix-*"
+//
+// The pattern syntax is:
+//
+//	pattern:
+//		{ term }
+//	term:
+//		'*'         matches any sequence of non-/ characters
+//		'?'         matches any single non-/ character
+//		'[' [ '^' ] { character-range } ']'
+//		            character class (must be non-empty)
+//		c           matches character c (c != '*', '?', '\\', '[')
+//		'\\' c      matches character c
+//
+//	character-range:
+//		c           matches character c (c != '\\', '-', ']')
+//		'\\' c      matches character c
+//		lo '-' hi   matches character c for lo <= c <= hi
+//
+// Keys requires pattern to match all of name, not just a substring.
 //
 // If cache is disabled (see [Disabled]), this function returns [ErrDisabled].
 func Keys(pattern string) (keys []string, err error) {
+	if err := validate(pattern); err != nil {
+		return nil, err
+	}
+
+	if _, err := path.Match(pattern, ""); err != nil {
+		return nil, err
+	}
+
 	dir, err := cacheDir()
 	if err != nil {
 		return nil, err
@@ -68,13 +105,71 @@ func Keys(pattern string) (keys []string, err error) {
 	kvMu.RLock()
 	defer kvMu.RUnlock()
 
-	return fs.Glob(os.DirFS(dir), pattern)
+	entries, err := os.ReadDir(dir)
+	if err != nil {
+		return nil, err
+	}
+
+	keys = make([]string, 0, len(keys))
+
+	for _, e := range entries {
+		key, err := extractKey(filepath.Join(dir, e.Name()))
+		if err != nil {
+			return nil, err
+		}
+
+		if ok, _ := path.Match(pattern, key); ok {
+			keys = append(keys, key)
+		}
+	}
+
+	return keys, nil
+}
+
+func extractKey(p string) (string, error) {
+	f, err := os.Open(p)
+	if err != nil {
+		return "", err
+	}
+	defer f.Close()
+
+	var collected []byte
+
+	// heuristic
+	chunk := make([]byte, 32)
+
+	for {
+		n, err := io.ReadAtLeast(f, chunk, 1)
+		if err != nil {
+			if errors.Is(err, io.EOF) {
+				return "", nil
+			}
+
+			return "", err
+		}
+
+		before, _, ok := bytes.Cut(chunk[:n], []byte{0})
+		if ok {
+			if len(collected) == 0 {
+				return string(before), nil
+			}
+
+			return string(append(collected, before...)), nil
+		}
+
+		collected = append(collected, before...)
+	}
 }
 
 // Get cached object by the given key.
+// Key must not contain a NUL-byte.
 //
 // If cache is disabled (see [Disabled]), this function returns [ErrDisabled].
 func Get(key string) ([]byte, error) {
+	if err := validate(key); err != nil {
+		return nil, err
+	}
+
 	dir, err := cacheDir()
 	if err != nil {
 		return nil, err
@@ -83,15 +178,40 @@ func Get(key string) ([]byte, error) {
 	kvMu.RLock()
 	defer kvMu.RUnlock()
 
-	path := filepath.Join(dir, sanitizeFilename(key))
+	h, err := hash(key)
+	if err != nil {
+		return nil, err
+	}
 
-	return os.ReadFile(path)
+	p := filepath.Join(dir, h)
+
+	_, err = os.Stat(p)
+	if err != nil {
+		return nil, ErrNotFound
+	}
+
+	value, err := os.ReadFile(p)
+	if err != nil {
+		return nil, err
+	}
+
+	_, after, ok := bytes.Cut(value, []byte{0})
+	if !ok {
+		return value, nil
+	}
+
+	return after, nil
 }
 
 // Set saves value to cache with the given key.
+// Key must not contain a NUL-byte.
 //
 // If cache is disabled (see [Disabled]), this function returns [ErrDisabled].
 func Set(key string, value []byte) error {
+	if err := validate(key); err != nil {
+		return err
+	}
+
 	dir, err := cacheDir()
 	if err != nil {
 		return err
@@ -100,15 +220,32 @@ func Set(key string, value []byte) error {
 	kvMu.Lock()
 	defer kvMu.Unlock()
 
-	path := filepath.Join(dir, sanitizeFilename(key))
+	h, err := hash(key)
+	if err != nil {
+		return err
+	}
 
-	return os.WriteFile(path, value, permFile)
+	p := filepath.Join(dir, h)
+
+	buf := bytes.NewBufferString(key)
+
+	buf.Grow(1 + len(value))
+
+	buf.WriteByte(0)
+	buf.Write(value)
+
+	return os.WriteFile(p, buf.Bytes(), permFile)
 }
 
 // Remove object from cache by the given key.
+// Key must not contain a NUL-byte.
 //
 // If cache is disabled (see [Disabled]), this function returns [ErrDisabled].
 func Remove(key string) error {
+	if err := validate(key); err != nil {
+		return err
+	}
+
 	dir, err := cacheDir()
 	if err != nil {
 		return err
@@ -117,9 +254,14 @@ func Remove(key string) error {
 	kvMu.Lock()
 	defer kvMu.Unlock()
 
-	path := filepath.Join(dir, sanitizeFilename(key))
+	h, err := hash(key)
+	if err != nil {
+		return err
+	}
 
-	return os.Remove(path)
+	p := filepath.Join(dir, h)
+
+	return os.Remove(p)
 }
 
 func cacheDir() (string, error) {
@@ -146,25 +288,21 @@ func parseBool(s string) bool {
 	return b
 }
 
-func sanitizeFilename(name string) string {
-	var sb strings.Builder
-
-	sb.Grow(len(name))
-
-	const (
-		invalid     = `\/<>:\"|?*.`
-		replacement = '-'
-	)
-
-	for _, r := range name {
-		switch {
-		case r == 0, unicode.IsControl(r), strings.ContainsRune(invalid, r):
-			sb.WriteRune(replacement)
-
-		default:
-			sb.WriteRune(r)
-		}
+func validate(key string) error {
+	if slices.Contains([]byte(key), 0) {
+		return ErrInvalidKey
 	}
 
-	return sb.String()
+	return nil
+}
+
+func hash(key string) (string, error) {
+	h := fnv.New64a()
+
+	_, err := h.Write([]byte(key))
+	if err != nil {
+		return "", err
+	}
+
+	return strconv.FormatUint(h.Sum64(), 36), nil
 }