+++ testocache/cache_test.go
@@ -1,11 +1,45 @@
 package testocache
 
 import (
+	"errors"
 	"slices"
 	"testing"
 )
 
-func Test(t *testing.T) {
+func TestInvalidKey(t *testing.T) {
+	t.Parallel()
+
+	const invalid = "foo\x00bar"
+
+	t.Run("set", func(t *testing.T) {
+		t.Parallel()
+
+		err := Set(invalid, []byte("..."))
+		if !errors.Is(err, ErrInvalidKey) {
+			t.Fatalf("err is not ErrInvalidKey: %v", err)
+		}
+	})
+
+	t.Run("get", func(t *testing.T) {
+		t.Parallel()
+
+		_, err := Get(invalid)
+		if !errors.Is(err, ErrInvalidKey) {
+			t.Fatalf("err is not ErrInvalidKey: %v", err)
+		}
+	})
+
+	t.Run("remove", func(t *testing.T) {
+		t.Parallel()
+
+		err := Remove(invalid)
+		if !errors.Is(err, ErrInvalidKey) {
+			t.Fatalf("err is not ErrInvalidKey: %v", err)
+		}
+	})
+}
+
+func TestFlow(t *testing.T) {
 	t.Parallel()
 
 	for _, tt := range []struct {
@@ -13,7 +47,7 @@ func Test(t *testing.T) {
 		Value string
 	}{
 		{Key: "my-key", Value: "lorem ipsum\ndolor sit \t\tamet"},
-		{Key: "key/with/slash", Value: "other value"},
+		{Key: "key~with~tilde", Value: "other value"},
 	} {
 		t.Run("with key: "+tt.Key, func(t *testing.T) {
 			err := Set(tt.Key, []byte(tt.Value))
@@ -37,7 +71,9 @@ func Test(t *testing.T) {
 		t.Fatalf("failed to get keys: %v", err)
 	}
 
-	wantKeys := []string{"key-with-slash", "my-key"}
+	slices.Sort(keys)
+
+	wantKeys := []string{"key~with~tilde", "my-key"}
 	if !slices.Equal(keys, wantKeys) {
 		t.Fatalf("keys: want %v, got %v", wantKeys, keys)
 	}
@@ -47,5 +83,10 @@ func Test(t *testing.T) {
 		if err != nil {
 			t.Errorf("remove key %q: %v", k, err)
 		}
+	}
+
+	_, err = Get("unknown-key")
+	if !errors.Is(err, ErrNotFound) {
+		t.Fatal("expected not found error")
 	}
 }